split action creators tests on different modules and solved issues created by it
